# Conflicts:
#	public-test-module/src/test/java/xmu/oomall/publictest/goods/GoodsIdTest.java

+++ goods-module/src/main/java/xmu/oomall/controller/GoodsController.java
@@ -125,6 +125,10 @@ public class GoodsController {
     @DeleteMapping("/brands/{id}")
     @ApiOperation(value = "删除一个品牌/delete")
     public Object deleteBrand(HttpServletRequest request,@PathVariable Integer id){
+        String token=request.getHeader("authorization");
+        if(token==null) {
+            return ResponseUtil.fail(660,"用户无操作权限");
+        }
         int ret= brandService.deleteBrand(id);
         if(ret==0){
             Log log=createLog(request, 0, 0, "删除品牌",id);
